refactor: delete unnecessary files
