+++ react-components/src/pages/Forms/index.tsx
@@ -4,7 +4,6 @@ import FormData from './components/FormData/index';
 import { BeastCard } from '../../types/beastCard';
 import CardForm from './components/CardForm';
 import Popup from './../../components/Popup/index';
-import FormData2 from './components/FormData/index copy';
 
 const Forms = () => {
   const [cards, setCards] = useState<BeastCard[]>([]);
@@ -24,7 +23,7 @@ const Forms = () => {
 
   return (
     <div className="main">
-      <FormData2 handleCard={handleCard} cards={cards.length} />
+      <FormData handleCard={handleCard} cards={cards.length} />
       <div className="forms__cards">
         <CardForm cards={cards} />
       </div>
